Fix CI compilation and workflow configuration issues
- Remove -Wconversion/-Wsign-conversion flags (false positives with PQC reference code)
- Add explicit casts in reduce.c, poly.c, mlkem.c for GCC/Clang warnings
- Fix null-dereference warning in decaps (move pointer calc after null check)
- Fix test binary names in all workflows (test_kem_all -> test_kem, test_sig_all -> test_sig)
- Fix benchmark binary path (build/bench_all -> build/tests/bench_all)
- Add missing cmake/ninja-build dependency installation steps in CI jobs

+++ src/core/kem/mlkem/poly.c
@@ -426,7 +426,7 @@ void pqc_mlkem_poly_basemul_montgomery(pqc_mlkem_poly *r,
 void pqc_mlkem_poly_tomont(pqc_mlkem_poly *r)
 {
     unsigned int i;
-    const int16_t f = (1ULL << 32) % PQC_MLKEM_Q;
+    const int16_t f = (int16_t)((1ULL << 32) % PQC_MLKEM_Q);
     for (i = 0; i < PQC_MLKEM_N; i++)
         r->coeffs[i] = pqc_mlkem_montgomery_reduce((int32_t)r->coeffs[i] * f);
 }
